requirements.txt
python-dotenv
streamlit
requests
cassandra-driver
astrapy
python-dotenv
streamlit
requests
matplotlib
